+++ src/main/java/com/aleksander/restaurant/reservation/dto/stats/ReservationCountDTO.java
@@ -1,0 +1,11 @@
+package com.aleksander.restaurant.reservation.dto.stats;
+
+import lombok.Builder;
+import lombok.Data;
+
+@Data
+@Builder
+public class ReservationCountDTO {
+    private int days;
+    private long reservationsCount;
+}